// src/app/api/api/models/operador-logistico-response.ts
/* tslint:disable */
/* eslint-disable */
export interface OperadorLogisticoResponse {
  activo?: boolean;
  contacto?: string;
  correo?: string;
  direccion?: string;
  estadoDsc?: string;
  nombreComercial?: string;
  operadorLogisticoId?: number;
  razonSocial?: string;
  ruc?: string;
  telefono?: string;
}